+++ protected/humhub/docs/CHANGELOG.md
@@ -21,6 +21,8 @@ HumHub Change Log
 - Enh: Toggle comment box
 - Enh: Added global copyToClipboard
 - Enh: Added "weekly" mail summary interval
+- Fix: Invalid temp.css file
+- Fix: Default stream sort setting not applied
 
 1.2.0-beta.3 (March 20, 2017)
 --------------------------------